+++ src/OpenTelemetry/CHANGELOG.md
@@ -6,6 +6,9 @@
   `FormatException` if it fails to parse any of the supported environment
   variables.
 
+* Added `BaseExporter.ForceFlush`.
+  ([#2525](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2525))
+
 ## 1.2.0-beta1
 
 Released 2021-Oct-08